core: lock/release of verifiers and users.
Verifiers lock first unverified items, others can't modify it in the mean time.

Once the verification is done, it sends the new value and all verifiers release the lock.

Users also receive the new value and update their local DB. Users are not part of the lock&release: it is supposed users use all available data, verified or not. Yes it would be possible for users to only process items that are verified, as by default, received items have status unverified.

+++ main.go
@@ -23,7 +23,7 @@ func main() {
 		// node.Sensor(node_name)
 
 		interval := time.Duration(2) * time.Second
-		errorRate := 0.1
+		errorRate := float32(0.1)
 
 		child_node = node.NewSensorNode(*node_id, interval, errorRate)
 	case "verifier":
